+++ CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to this project will be documented in this file.
 - New Bloblang methods `encrypt_aes` and `decrypt_aes` added.
 - New field `static_headers` added to the `kafka` output.
 - New field `enabled` added to the `http` config section.
+- Experimental CLI flag `-resources` added for specifying files containing extra
+  resources.
 
 ### Fixed